worker/requirements.txt
# Core dependencies for the worker runtime.
requests
serpapi
python-dotenv
psycopg2-binary
flask
googlemaps
geopy
beautifulsoup4
phonenumbers
playwright
